requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-login==0.6.2
flask-jwt-extended==4.4.4
flask-cors==3.0.10
psycopg2-binary==2.9.5
python-dotenv==1.0.0
werkzeug==2.2.3